Refactor GetSkdeParams by moving shared types and handler to common module
