+++ src/rpc/cluster/mod.rs
@@ -1,5 +1,4 @@
 mod get_key_generator_list;
-mod get_skde_params;
 mod request_submit_partial_key;
 mod sync_finalized_partial_keys;
 mod sync_key_generator;
@@ -10,7 +9,6 @@ mod submit_partial_key;
 mod sync_decryption_key;
 
 pub use get_key_generator_list::*;
-pub use get_skde_params::*;
 pub use request_submit_partial_key::*;
 pub use submit_final_reveal::*;
 pub use submit_partial_key::*;
