+++ src/rpc/solver/mod.rs
@@ -1,7 +1,5 @@
-mod get_skde_params;
 mod submit_decryption_key;
 mod sync_finalized_partial_keys;
 
-pub use get_skde_params::*;
 pub use submit_decryption_key::*;
 pub use sync_finalized_partial_keys::*;
